refactor(state): split Inspector to remove internal/graph from cmd/bud deps
Move all graph.DB-dependent methods from state.Inspector into a new
internal/state/graphdb.Inspector type. This breaks internal/graph out
of cmd/bud's transitive dependency tree.

- state.Inspector: file-based state only (percepts, threads, queues, etc.)
- state/graphdb.Inspector: embeds state.Inspector, adds graph DB methods
- cmd/bud uses state.NewInspector(statePath) — no graph dep
- cmd/bud-state uses graphdb.NewInspector(statePath, graphDB) — graph dep isolated
