+++ internal/state/inspect.go
@@ -9,31 +9,18 @@ import (
 	"strings"
 	"time"
 
-	"github.com/vthunder/bud2/internal/graph"
-	"github.com/vthunder/bud2/internal/profiling"
 	"github.com/vthunder/bud2/internal/types"
 )
 
-// Embedder generates embeddings for text
-type Embedder interface {
-	Embed(text string) ([]float64, error)
-}
-
-// Inspector provides state introspection capabilities
+// Inspector provides state introspection capabilities for file-based state.
+// For graph database access, use GraphInspector.
 type Inspector struct {
 	statePath string
-	graphDB   *graph.DB
-	embedder  Embedder
 }
 
 // NewInspector creates a new state inspector
-func NewInspector(statePath string, graphDB *graph.DB) *Inspector {
-	return &Inspector{statePath: statePath, graphDB: graphDB}
-}
-
-// SetEmbedder sets the embedder for memory search (optional)
-func (i *Inspector) SetEmbedder(e Embedder) {
-	i.embedder = e
+func NewInspector(statePath string) *Inspector {
+	return &Inspector{statePath: statePath}
 }
 
 // ComponentSummary holds summary for one state component
@@ -62,14 +49,6 @@ type HealthReport struct {
 // Summary returns a summary of all state components
 func (i *Inspector) Summary() (*StateSummary, error) {
 	summary := &StateSummary{}
-
-	// Count traces from graph DB
-	if i.graphDB != nil {
-		total, err := i.graphDB.CountTraces()
-		if err == nil {
-			summary.Traces.Total = total
-		}
-	}
 
 	// Count percepts
 	percepts, err := i.loadPercepts()
@@ -119,366 +98,6 @@ func (i *Inspector) Health() (*HealthReport, error) {
 	}
 
 	return report, nil
-}
-
-// TraceSummary is a condensed view of a trace
-type TraceSummary struct {
-	ID        string    `json:"id"`
-	Content   string    `json:"content"`
-	Strength  int       `json:"strength"`
-	CreatedAt time.Time `json:"created_at"`
-}
-
-// ListTraces returns summaries of all traces
-func (i *Inspector) ListTraces() ([]TraceSummary, error) {
-	if i.graphDB == nil {
-		return nil, fmt.Errorf("graph database not initialized")
-	}
-
-	traces, err := i.graphDB.GetAllTraces()
-	if err != nil {
-		return nil, err
-	}
-
-	result := make([]TraceSummary, 0, len(traces))
-	for _, t := range traces {
-		content := t.Summary
-		if len(content) > 100 {
-			content = content[:100] + "..."
-		}
-		result = append(result, TraceSummary{
-			ID:        t.ID,
-			Content:   content,
-			Strength:  t.Strength,
-			CreatedAt: t.CreatedAt,
-		})
-	}
-	return result, nil
-}
-
-// GetTrace returns full trace by ID
-func (i *Inspector) GetTrace(id string) (*types.Trace, error) {
-	if i.graphDB == nil {
-		return nil, fmt.Errorf("graph database not initialized")
-	}
-
-	gt, err := i.graphDB.GetTrace(id)
-	if err != nil {
-		return nil, err
-	}
-	if gt == nil {
-		return nil, fmt.Errorf("trace not found: %s", id)
-	}
-
-	// Convert graph.Trace to types.Trace
-	return &types.Trace{
-		ID:         gt.ID,
-		Content:    gt.Summary,
-		Embedding:  gt.Embedding,
-		Activation: gt.Activation,
-		Strength:   gt.Strength,
-		CreatedAt:  gt.CreatedAt,
-		LastAccess: gt.LastAccessed,
-	}, nil
-}
-
-// DeleteTrace removes a trace by ID
-func (i *Inspector) DeleteTrace(id string) error {
-	if i.graphDB == nil {
-		return fmt.Errorf("graph database not initialized")
-	}
-	return i.graphDB.DeleteTrace(id)
-}
-
-// EpisodeSummary is a condensed view of an episode
-type EpisodeSummary struct {
-	ID        string    `json:"id"`
-	Content   string    `json:"content"`
-	Source    string    `json:"source"`
-	Author    string    `json:"author,omitempty"`
-	Channel   string    `json:"channel,omitempty"`
-	Timestamp time.Time `json:"timestamp"`
-}
-
-// ListEpisodes returns summaries of recent episodes
-func (i *Inspector) ListEpisodes(limit int) ([]EpisodeSummary, error) {
-	if i.graphDB == nil {
-		return nil, fmt.Errorf("graph database not initialized")
-	}
-
-	episodes, err := i.graphDB.GetAllEpisodes(limit)
-	if err != nil {
-		return nil, err
-	}
-
-	result := make([]EpisodeSummary, 0, len(episodes))
-	for _, ep := range episodes {
-		content := ep.Content
-		if len(content) > 100 {
-			content = content[:100] + "..."
-		}
-		result = append(result, EpisodeSummary{
-			ID:        ep.ID,
-			Content:   content,
-			Source:    ep.Source,
-			Author:    ep.Author,
-			Channel:   ep.Channel,
-			Timestamp: ep.TimestampEvent,
-		})
-	}
-	return result, nil
-}
-
-// GetEpisode returns full episode by ID
-func (i *Inspector) GetEpisode(id string) (*graph.Episode, error) {
-	if i.graphDB == nil {
-		return nil, fmt.Errorf("graph database not initialized")
-	}
-	return i.graphDB.GetEpisode(id)
-}
-
-// CountEpisodes returns the total number of episodes
-func (i *Inspector) CountEpisodes() (int, error) {
-	if i.graphDB == nil {
-		return 0, fmt.Errorf("graph database not initialized")
-	}
-	return i.graphDB.CountEpisodes()
-}
-
-// EntitySummary is a condensed view of an entity
-type EntitySummary struct {
-	ID       string   `json:"id"`
-	Name     string   `json:"name"`
-	Type     string   `json:"type"`
-	Salience float64  `json:"salience"`
-	Aliases  []string `json:"aliases,omitempty"`
-}
-
-// ListEntities returns summaries of entities ordered by salience
-func (i *Inspector) ListEntities(limit int) ([]EntitySummary, error) {
-	if i.graphDB == nil {
-		return nil, fmt.Errorf("graph database not initialized")
-	}
-
-	entities, err := i.graphDB.GetAllEntities(limit)
-	if err != nil {
-		return nil, err
-	}
-
-	result := make([]EntitySummary, 0, len(entities))
-	for _, e := range entities {
-		result = append(result, EntitySummary{
-			ID:       e.ID,
-			Name:     e.Name,
-			Type:     string(e.Type),
-			Salience: e.Salience,
-			Aliases:  e.Aliases,
-		})
-	}
-	return result, nil
-}
-
-// GetEntity returns full entity by ID
-func (i *Inspector) GetEntity(id string) (*graph.Entity, error) {
-	if i.graphDB == nil {
-		return nil, fmt.Errorf("graph database not initialized")
-	}
-	return i.graphDB.GetEntity(id)
-}
-
-// CountEntities returns the total number of entities
-func (i *Inspector) CountEntities() (int, error) {
-	if i.graphDB == nil {
-		return 0, fmt.Errorf("graph database not initialized")
-	}
-	return i.graphDB.CountEntities()
-}
-
-// NodeInfo holds information about a node and its relationships
-type NodeInfo struct {
-	ID       string            `json:"id"`
-	Type     string            `json:"type"` // "trace", "episode", "entity"
-	Data     any               `json:"data"` // The actual node data
-	Links    map[string][]Link `json:"links,omitempty"`
-}
-
-// Link represents a relationship to another node
-type Link struct {
-	ID       string  `json:"id"`
-	Type     string  `json:"type,omitempty"`
-	Weight   float64 `json:"weight,omitempty"`
-	Preview  string  `json:"preview,omitempty"`
-}
-
-// GetNodeInfo returns information about a node and its relationships
-func (i *Inspector) GetNodeInfo(id string) (*NodeInfo, error) {
-	if i.graphDB == nil {
-		return nil, fmt.Errorf("graph database not initialized")
-	}
-
-	// Try to find the node in each table
-	// Check traces first
-	if trace, _ := i.graphDB.GetTrace(id); trace != nil {
-		info := &NodeInfo{
-			ID:   id,
-			Type: "trace",
-			Data: trace,
-			Links: make(map[string][]Link),
-		}
-
-		// Get source episodes
-		if sources, _ := i.graphDB.GetTraceSources(id); len(sources) > 0 {
-			var links []Link
-			for _, srcID := range sources {
-				link := Link{ID: srcID}
-				if ep, _ := i.graphDB.GetEpisode(srcID); ep != nil {
-					preview := ep.Content
-					if len(preview) > 50 {
-						preview = preview[:50] + "..."
-					}
-					link.Preview = preview
-				}
-				links = append(links, link)
-			}
-			info.Links["source_episodes"] = links
-		}
-
-		// Get linked entities
-		if entityIDs, _ := i.graphDB.GetTraceEntities(id); len(entityIDs) > 0 {
-			var links []Link
-			for _, entID := range entityIDs {
-				link := Link{ID: entID}
-				if ent, _ := i.graphDB.GetEntity(entID); ent != nil {
-					link.Preview = ent.Name
-					link.Type = string(ent.Type)
-				}
-				links = append(links, link)
-			}
-			info.Links["entities"] = links
-		}
-
-		// Get related traces
-		if neighbors, _ := i.graphDB.GetTraceNeighbors(id); len(neighbors) > 0 {
-			var links []Link
-			for _, n := range neighbors {
-				link := Link{ID: n.ID, Type: string(n.Type), Weight: n.Weight}
-				if t, _ := i.graphDB.GetTrace(n.ID); t != nil {
-					preview := t.Summary
-					if len(preview) > 50 {
-						preview = preview[:50] + "..."
-					}
-					link.Preview = preview
-				}
-				links = append(links, link)
-			}
-			info.Links["related_traces"] = links
-		}
-
-		return info, nil
-	}
-
-	// Check episodes
-	if episode, _ := i.graphDB.GetEpisode(id); episode != nil {
-		info := &NodeInfo{
-			ID:   id,
-			Type: "episode",
-			Data: episode,
-			Links: make(map[string][]Link),
-		}
-
-		// Get mentioned entities
-		if entities, _ := i.graphDB.GetEntitiesForEpisode(id); len(entities) > 0 {
-			var links []Link
-			for _, e := range entities {
-				links = append(links, Link{
-					ID:      e.ID,
-					Preview: e.Name,
-					Type:    string(e.Type),
-				})
-			}
-			info.Links["mentions"] = links
-		}
-
-		// Get episode neighbors (replies)
-		if neighbors, _ := i.graphDB.GetEpisodeNeighbors(id); len(neighbors) > 0 {
-			var links []Link
-			for _, n := range neighbors {
-				link := Link{ID: n.ID, Type: string(n.Type), Weight: n.Weight}
-				if ep, _ := i.graphDB.GetEpisode(n.ID); ep != nil {
-					preview := ep.Content
-					if len(preview) > 50 {
-						preview = preview[:50] + "..."
-					}
-					link.Preview = preview
-				}
-				links = append(links, link)
-			}
-			info.Links["replies"] = links
-		}
-
-		return info, nil
-	}
-
-	// Check entities
-	if entity, _ := i.graphDB.GetEntity(id); entity != nil {
-		info := &NodeInfo{
-			ID:   id,
-			Type: "entity",
-			Data: entity,
-			Links: make(map[string][]Link),
-		}
-
-		// Get episodes mentioning this entity
-		if episodeIDs, _ := i.graphDB.GetEpisodesForEntity(id); len(episodeIDs) > 0 {
-			var links []Link
-			for _, epID := range episodeIDs {
-				link := Link{ID: epID}
-				if ep, _ := i.graphDB.GetEpisode(epID); ep != nil {
-					preview := ep.Content
-					if len(preview) > 50 {
-						preview = preview[:50] + "..."
-					}
-					link.Preview = preview
-				}
-				links = append(links, link)
-			}
-			info.Links["mentioned_in"] = links
-		}
-
-		// Get traces involving this entity
-		if traceIDs, _ := i.graphDB.GetTracesForEntity(id); len(traceIDs) > 0 {
-			var links []Link
-			for _, tID := range traceIDs {
-				link := Link{ID: tID}
-				if t, _ := i.graphDB.GetTrace(tID); t != nil {
-					preview := t.Summary
-					if len(preview) > 50 {
-						preview = preview[:50] + "..."
-					}
-					link.Preview = preview
-				}
-				links = append(links, link)
-			}
-			info.Links["in_traces"] = links
-		}
-
-		// Get related entities
-		if neighbors, _ := i.graphDB.GetEntityRelations(id); len(neighbors) > 0 {
-			var links []Link
-			for _, n := range neighbors {
-				link := Link{ID: n.ID, Type: string(n.Type), Weight: n.Weight}
-				if e, _ := i.graphDB.GetEntity(n.ID); e != nil {
-					link.Preview = e.Name
-				}
-				links = append(links, link)
-			}
-			info.Links["related_entities"] = links
-		}
-
-		return info, nil
-	}
-
-	return nil, fmt.Errorf("node not found: %s", id)
 }
 
 // PerceptSummary is a condensed view of a percept
@@ -705,270 +324,6 @@ func (i *Inspector) ListSessions() ([]SessionInfo, error) {
 func (i *Inspector) ClearSessions() error {
 	path := filepath.Join(i.statePath, "system", "sessions.json")
 	return os.WriteFile(path, []byte("{}"), 0644)
-}
-
-
-// SearchResult represents a memory search result
-type SearchResult struct {
-	ID         string    `json:"id"`
-	Summary    string    `json:"summary"`
-	Activation float64   `json:"activation"`
-	CreatedAt  time.Time `json:"created_at"`
-}
-
-// SearchMemory searches long-term memory for traces matching the query.
-// If useContext is true, also considers currently-activated traces as additional seeds,
-// biasing results toward memories connected to the current working context.
-func (i *Inspector) SearchMemory(query string, limit int) ([]SearchResult, error) {
-	return i.SearchMemoryWithContext(query, limit, true)
-}
-
-// SearchMemoryWithContext searches memory with optional context awareness.
-func (i *Inspector) SearchMemoryWithContext(query string, limit int, useContext bool) ([]SearchResult, error) {
-	// Track total search time
-	defer profiling.Get().StartWithMetadata("search_memory", "memory_search.total", map[string]interface{}{
-		"query_length": len(query),
-		"limit":        limit,
-		"use_context":  useContext,
-	})()
-
-	if i.graphDB == nil {
-		return nil, fmt.Errorf("graph database not initialized")
-	}
-	if i.embedder == nil {
-		return nil, fmt.Errorf("embedder not configured - cannot search memory")
-	}
-
-	// Default limit
-	if limit <= 0 {
-		limit = 5
-	}
-	if limit > 20 {
-		limit = 20
-	}
-
-	// Generate embedding for query (track timing)
-	var queryEmb []float64
-	{
-		stopEmbedding := profiling.Get().Start("search_memory", "memory_search.embedding")
-		var err error
-		queryEmb, err = i.embedder.Embed(query)
-		stopEmbedding()
-		if err != nil {
-			return nil, fmt.Errorf("failed to embed query: %w", err)
-		}
-	}
-
-	var result *graph.RetrievalResult
-
-	if useContext {
-		// Get currently-activated traces to use as additional context
-		stopGetActivated := profiling.Get().Start("search_memory", "memory_search.get_activated")
-		activatedTraces, err := i.graphDB.GetActivatedTraces(0.3, 10)
-		stopGetActivated()
-		if err != nil {
-			// Fall back to regular retrieval if we can't get activated traces
-			stopRetrieve := profiling.Get().StartWithMetadata("search_memory", "memory_search.retrieve", map[string]interface{}{
-				"mode": "fallback",
-			})
-			result, err = i.graphDB.Retrieve(queryEmb, query, limit)
-			stopRetrieve()
-			if err != nil {
-				return nil, fmt.Errorf("retrieval failed: %w", err)
-			}
-		} else {
-			// Extract IDs from activated traces
-			contextIDs := make([]string, 0, len(activatedTraces))
-			for _, t := range activatedTraces {
-				contextIDs = append(contextIDs, t.ID)
-			}
-			// Use context-aware retrieval
-			stopRetrieve := profiling.Get().StartWithMetadata("search_memory", "memory_search.retrieve_with_context", map[string]interface{}{
-				"context_count": len(contextIDs),
-			})
-			var err error
-			result, err = i.graphDB.RetrieveWithContext(queryEmb, query, contextIDs, limit)
-			stopRetrieve()
-			if err != nil {
-				return nil, fmt.Errorf("context retrieval failed: %w", err)
-			}
-		}
-	} else {
-		// Use standard retrieval without context
-		stopRetrieve := profiling.Get().StartWithMetadata("search_memory", "memory_search.retrieve", map[string]interface{}{
-			"mode": "standard",
-		})
-		var err error
-		result, err = i.graphDB.Retrieve(queryEmb, query, limit)
-		stopRetrieve()
-		if err != nil {
-			return nil, fmt.Errorf("retrieval failed: %w", err)
-		}
-	}
-
-	// Convert to SearchResult (track timing for result processing)
-	stopFormat := profiling.Get().StartWithMetadata("search_memory", "memory_search.format_results", map[string]interface{}{
-		"result_count": len(result.Traces),
-	})
-	var results []SearchResult
-	for _, t := range result.Traces {
-		results = append(results, SearchResult{
-			ID:         t.ID,
-			Summary:    t.Summary,
-			Activation: t.Activation,
-			CreatedAt:  t.CreatedAt,
-		})
-	}
-	stopFormat()
-
-	return results, nil
-}
-
-// TraceContext provides detailed context for a trace including source episodes and linked entities
-type TraceContext struct {
-	Trace    TraceSummary     `json:"trace"`
-	Episodes []EpisodeSummary `json:"source_episodes"`
-	Entities []EntitySummary  `json:"linked_entities"`
-}
-
-// GetTraceContext retrieves detailed context for a trace
-func (i *Inspector) GetTraceContext(traceID string) (*TraceContext, error) {
-	if i.graphDB == nil {
-		return nil, fmt.Errorf("graph database not initialized")
-	}
-
-	// Get the trace
-	trace, err := i.graphDB.GetTrace(traceID)
-	if err != nil {
-		return nil, err
-	}
-	if trace == nil {
-		return nil, fmt.Errorf("trace not found: %s", traceID)
-	}
-
-	ctx := &TraceContext{
-		Trace: TraceSummary{
-			ID:        trace.ID,
-			Content:   trace.Summary,
-			Strength:  trace.Strength,
-			CreatedAt: trace.CreatedAt,
-		},
-	}
-
-	// Get source episodes
-	episodeIDs, err := i.graphDB.GetTraceSources(traceID)
-	if err == nil {
-		for _, epID := range episodeIDs {
-			ep, err := i.graphDB.GetEpisode(epID)
-			if err != nil || ep == nil {
-				continue
-			}
-			content := ep.Content
-			if len(content) > 200 {
-				content = content[:200] + "..."
-			}
-			ctx.Episodes = append(ctx.Episodes, EpisodeSummary{
-				ID:        ep.ID,
-				Content:   content,
-				Source:    ep.Source,
-				Author:    ep.Author,
-				Channel:   ep.Channel,
-				Timestamp: ep.TimestampEvent,
-			})
-		}
-	}
-
-	// Get linked entities
-	entityIDs, err := i.graphDB.GetTraceEntities(traceID)
-	if err == nil {
-		for _, entID := range entityIDs {
-			ent, err := i.graphDB.GetEntity(entID)
-			if err != nil || ent == nil {
-				continue
-			}
-			ctx.Entities = append(ctx.Entities, EntitySummary{
-				ID:       ent.ID,
-				Name:     ent.Name,
-				Type:     string(ent.Type),
-				Salience: ent.Salience,
-				Aliases:  ent.Aliases,
-			})
-		}
-	}
-
-	return ctx, nil
-}
-
-// QueryTrace runs a question against the source episodes of a trace using the LLM.
-// level parameter controls compression: 0=raw episodes, 1=L1 summary (default), 2=L2 summary
-func (i *Inspector) QueryTrace(traceID, question string, level int) (string, error) {
-	if i.graphDB == nil {
-		return "", fmt.Errorf("graph database not initialized")
-	}
-
-	// Get source episodes for context
-	episodeIDs, err := i.graphDB.GetTraceSources(traceID)
-	if err != nil {
-		return "", fmt.Errorf("failed to get trace sources: %w", err)
-	}
-
-	if len(episodeIDs) == 0 {
-		return "", fmt.Errorf("trace has no source episodes")
-	}
-
-	// Validate compression level
-	if level < 0 || level > 2 {
-		return "", fmt.Errorf("invalid compression level %d (must be 0-2)", level)
-	}
-
-	// Gather episode content (raw or compressed)
-	var context strings.Builder
-	var usedCompression bool
-	for _, epID := range episodeIDs {
-		ep, err := i.graphDB.GetEpisode(epID)
-		if err != nil || ep == nil {
-			continue
-		}
-
-		var content string
-		if level > 0 {
-			// Try to get compressed summary
-			summary, err := i.graphDB.GetEpisodeSummary(epID, level)
-			if err == nil && summary != nil {
-				content = summary.Summary
-				usedCompression = true
-			} else {
-				// Fall back to raw content if summary not available
-				content = ep.Content
-			}
-		} else {
-			// Level 0: use raw content
-			content = ep.Content
-		}
-
-		context.WriteString(fmt.Sprintf("[%s] %s: %s\n",
-			ep.TimestampEvent.Format("2006-01-02 15:04"), ep.Author, content))
-	}
-
-	if context.Len() == 0 {
-		return "", fmt.Errorf("no episode content available")
-	}
-
-	// Build result string
-	var result strings.Builder
-	result.WriteString(fmt.Sprintf("Context from %d source episodes", len(episodeIDs)))
-	if usedCompression {
-		result.WriteString(fmt.Sprintf(" (compression level %d)", level))
-	}
-	result.WriteString(":\n\n")
-	result.WriteString(context.String())
-
-	if question != "" {
-		result.WriteString("\n(LLM query not yet implemented - add Generator interface to Inspector)\n")
-		result.WriteString(fmt.Sprintf("Question: %s\n", question))
-	}
-
-	return result.String(), nil
 }
 
 // Helper methods
